clean up logger usage
