+++ src/types.ts
@@ -1,6 +1,1 @@
-export type Logger = {
-  error: typeof console.log;
-  info: typeof console.log;
-  debug: typeof console.log;
-  warn: typeof console.log;
 };
